Exception handling

+++ src/main/java/com/devrenno/dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.devrenno.dscommerce.services;
 import com.devrenno.dscommerce.dto.ProductDTO;
 import com.devrenno.dscommerce.entities.Product;
 import com.devrenno.dscommerce.repositories.ProductRepository;
+import com.devrenno.dscommerce.services.exceptions.DatabaseException;
+import com.devrenno.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -17,7 +22,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findByid(Long id){
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ProductDTO(product);
     }
 
@@ -40,16 +46,28 @@ public class ProductService {
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO){
 
-        Product product = repository.getReferenceById(id);
-        copyDtoToEntity(productDTO, product);
-        product = repository.save(product);
+        try {
+            Product product = repository.getReferenceById(id);
+            copyDtoToEntity(productDTO, product);
+            product = repository.save(product);
 
-        return new ProductDTO(product);
+            return new ProductDTO(product);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void deleteByid(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)){
+            throw  new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw  new DatabaseException("falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO productDTO, Product product) {
